♻️ refactor: split viever into model and chains
